# specs/001-package-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Elsa Platform API
  version: 0.1.0
  description: Public and admin REST contracts for package, feature, compatibility, sync, validation, and approval workflows.
servers:
  - url: http://localhost:5000
security: []
...
paths:
  /api/packages:
    get:
      tags: [Public Packages]
      summary: List public packages
      parameters:
        - $ref: '#/components/parameters/SourceIdsQuery'
        - $ref: '#/components/parameters/PackageIdQuery'
        - $ref: '#/components/parameters/TagQuery'
        - $ref: '#/components/parameters/FeatureIdQuery'
        - $ref: '#/components/parameters/ElsaVersionQuery'
      responses:
        '200':
          description: Public packages containing only listed, approved, valid versions.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PackageSummary'
...
  /api/sources/{sourceId}/packages/{packageId}:
    get:
      tags: [Public Packages]
      summary: Get source-qualified public package details
      parameters:
        - $ref: '#/components/parameters/SourceIdPath'
        - $ref: '#/components/parameters/PackageIdPath'
      responses:
        '200':
          description: Public package details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageDetails'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/sources:
    get:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: List package sources
      responses:
        '200':
          description: Package sources.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminPackageSource'
    post:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: Create package source
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPackageSourceRequest'
      responses:
        '201':
          description: Created source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackageSource'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/sources/{id}:
    put:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: Update package source
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPackageSourceRequest'
      responses:
        '200':
          description: Updated source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackageSource'
    delete:
      tags: [Admin Sources]
      security:
        - ApiKeyAuth: []
      summary: Delete package source
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Source deleted.
  /api/admin/sync:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for all enabled sources
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
  /api/admin/sync/sources/{sourceId}:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for one source
      parameters:
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
  /api/admin/sync/packages/{packageId}:
    post:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Trigger sync for one package ID
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
      responses:
        '202':
          $ref: '#/components/responses/SyncAccepted'
  /api/admin/sync-runs:
    get:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: List sync runs
      responses:
        '200':
          description: Sync runs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncRunSummary'
  /api/admin/sync-runs/{id}:
    get:
      tags: [Admin Sync]
      security:
        - ApiKeyAuth: []
      summary: Get sync run details
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Sync run details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncRunDetails'
  /api/admin/packages:
    get:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: List packages including admin states
      responses:
        '200':
          description: Admin packages.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminPackageSummary'
  /api/admin/packages/{packageId}:
    get:
      tags: [Admin Packages]
      security:
        - ApiKeyAuth: []
      summary: Get admin package details
      parameters:
        - $ref: '#/components/parameters/PackageIdPath'
      responses:
        '200':
          description: Admin package details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPackageDetails'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    SourceIdPath:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PackageIdPath:
      name: packageId
      in: path
      required: true
      schema:
        type: string
...
    PackageIdQuery:
      name: packageId
      in: query
      schema:
        type: string
    TagQuery:
      name: tag
      in: query
      schema:
        type: string
    FeatureIdQuery:
      name: featureId
      in: query
      schema:
        type: string
    ElsaVersionQuery:
      name: elsaVersion
      in: query
      schema:
        type: string
    SourceIdsQuery:
      name: sourceIds
      in: query
      schema:
        type: array
        items:
          type: string
          format: uuid
      style: form
      explode: true
...
  responses:
    NotFound:
      description: Resource not found or not visible.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    SyncAccepted:
      description: Sync accepted.
      content:
        application/json:
          schema:
            type: object
            properties:
              syncRunId:
                type: string
              status:
                type: string
                enum: [Running]
...
  schemas:
...
    PackageSummary:
      type: object
      required: [packageId, displayName, latestVersion]
      properties:
        packageId:
          type: string
        source:
          $ref: '#/components/schemas/PackageSourceProvenance'
        displayName:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        latestVersion:
          type: string
    PackageDetails:
      allOf:
        - $ref: '#/components/schemas/PackageSummary'
        - type: object
          properties:
            versions:
              type: array
              items:
                $ref: '#/components/schemas/PackageVersionSummary'
            documentation:
              $ref: '#/components/schemas/Documentation'
            license:
              $ref: '#/components/schemas/License'
    PackageVersionSummary:
      type: object
      required: [packageId, version]
      properties:
        packageId:
          type: string
        version:
          type: string
        source:
          $ref: '#/components/schemas/PackageSourceProvenance'
        publishedAt:
          type: string
          format: date-time
        indexedAt:
          type: string
          format: date-time
        featureCount:
          type: integer
...
    PackageSourceProvenance:
      type: object
      required: [id, name, url]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
...
    License:
      type: object
      properties:
        expression:
          type: string
        url:
          type: string
        requiresAcceptance:
          type: boolean
    Documentation:
      type: object
      properties:
        readmeUrl:
          type: string
        projectUrl:
          type: string
        releaseNotesUrl:
          type: string
        configurationUrl:
          type: string
    AdminPackageSource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [NuGetFeed]
        url:
          type: string
        enabled:
          type: boolean
        includePatterns:
          type: array
          items:
            type: string
        excludePatterns:
          type: array
          items:
            type: string
        approvalPolicy:
          type: string
          enum: [AutoApprove, Manual]
        versionDiscoveryPolicy:
          $ref: '#/components/schemas/PackageSourceVersionDiscoveryPolicy'
        lastSyncedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpsertPackageSourceRequest:
      type: object
      required: [name, type, url, enabled, includePatterns, approvalPolicy]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [NuGetFeed]
        url:
          type: string
        enabled:
          type: boolean
        includePatterns:
          type: array
          items:
            type: string
        excludePatterns:
          type: array
          items:
            type: string
        approvalPolicy:
          type: string
          enum: [AutoApprove, Manual]
        versionDiscoveryPolicy:
          type: string
          enum: [AllVersions, LatestStable, LatestIncludingPrerelease, LatestPreview]
          default: AllVersions
    PackageSourceVersionDiscoveryPolicy:
      type: string
      enum: [AllVersions, LatestStable, LatestIncludingPrerelease, LatestPreview]
    SyncRequest:
      type: object
      properties:
        forceReindex:
          type: boolean
          default: false
    SyncRunSummary:
      type: object
      properties:
        id:
          type: string
        trigger:
          type: string
        status:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        summaryCounters:
          type: object
          additionalProperties:
            type: integer
    SyncRunDetails:
      allOf:
        - $ref: '#/components/schemas/SyncRunSummary'
        - type: object
          properties:
            error:
              type: string
            items:
              type: array
              items:
                type: object
                properties:
                  sourceId:
                    type: string
                  packageId:
                    type: string
                  version:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
                  error:
                    type: string
    AdminPackageSummary:
      type: object
      properties:
        packageId:
          type: string
        sourceId:
          type: string
        approved:
          type: boolean
        listed:
          type: boolean
        latestVersion:
          type: string
    AdminPackageDetails:
      allOf:
        - $ref: '#/components/schemas/AdminPackageSummary'
        - type: object
          properties:
            versions:
              type: array
              items:
                type: object
                properties:
                  version:
                    type: string
                  validationStatus:
                    type: string
                  approvalStatus:
                    type: string
                  isListed:
                    type: boolean
                  suspiciousChangeDetected:
                    type: boolean
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string